refactor: file instruction operations
